Update and rename aautilities.sh to aa-util.sh

+++ module/aa-util.sh
@@ -100,28 +100,31 @@ module_intro() {
 
 }
 
-init_logowl() {
-
+logowl_init() {
     LOG_DIR="$1"
 
-    if [ -z "$LOG_DIR" ]; then
-        logowl "Log dir is NOT ordered! (1)" "ERROR"
-        return 1
-    fi
-
-    if [ ! -d "$LOG_DIR" ]; then
-        logowl "Log dir $LOG_DIR does NOT exist"
-        mkdir -p "$LOG_DIR" || {
-            logowl "Failed to create $LOG_DIR (2)" "ERROR"
-            return 2
-        }
-        logowl "Created $LOG_DIR"
-    fi
+    [ -z "$LOG_DIR" ] && return 1
+    [ ! -d "$LOG_DIR" ] && mkdir -p "$LOG_DIR" && logowl "Created $LOG_DIR"
 
 }
 
-logowl() {
+logowl_clean() {
+    log_dir="$1"
+    files_max="$2"
+    
+    [ -z "$log_dir" ] || [ ! -d "$log_dir" ] && return 1
+    [ -z "$files_max" ] && files_max=30
 
+    files_count=$(ls -1 "$log_dir" | wc -l)
+    if [ "$files_count" -gt "$files_max" ]; then
+        ls -1t "$log_dir" | tail -n +$((files_max + 1)) | while read -r file; do
+            rm -f "$log_dir/$file"
+        done
+    fi
+    return 0
+}
+
+logowl() {
     LOG_MSG="$1"
     LOG_MSG_LEVEL="$2"
     LOG_MSG_PREFIX=""
@@ -129,21 +132,21 @@ logowl() {
     [ -z "$LOG_MSG" ] && return 1
 
     case "$LOG_MSG_LEVEL" in
-        "TIPS") LOG_MSG_PREFIX="* " ;;
+        "TIPS") LOG_MSG_PREFIX="> " ;;
         "WARN") LOG_MSG_PREFIX="- Warn: " ;;
         "ERROR") LOG_MSG_PREFIX="! ERROR: " ;;
         "FATAL") LOG_MSG_PREFIX="× FATAL: " ;;
-        "SPACE") LOG_MSG_PREFIX="  " ;;
-        "NONE") LOG_MSG_PREFIX="" ;;
+        " ") LOG_MSG_PREFIX="  " ;;
+        "-") LOG_MSG_PREFIX="" ;;
         *) LOG_MSG_PREFIX="- " ;;
     esac
 
     if [ -n "$LOG_FILE" ]; then
         if [ "$LOG_MSG_LEVEL" = "ERROR" ] || [ "$LOG_MSG_LEVEL" = "FATAL" ]; then
-            echo "----------------------------------------------------------------------" >> "$LOG_FILE"
+            echo "----------------------------------------" >> "$LOG_FILE"
             echo "${LOG_MSG_PREFIX}${LOG_MSG}" >> "$LOG_FILE"
-            echo "----------------------------------------------------------------------" >> "$LOG_FILE"
-        elif [ "$LOG_MSG_LEVEL" = "NONE" ]; then
+            echo "----------------------------------------" >> "$LOG_FILE"
+        elif [ "$LOG_MSG_LEVEL" = "-" ]; then
             echo "$LOG_MSG" >> "$LOG_FILE"
         else
             echo "${LOG_MSG_PREFIX}${LOG_MSG}" >> "$LOG_FILE"
@@ -151,10 +154,10 @@ logowl() {
     else
         if command -v ui_print >/dev/null 2>&1; then
             if [ "$LOG_MSG_LEVEL" = "ERROR" ] || [ "$LOG_MSG_LEVEL" = "FATAL" ]; then
-                ui_print "----------------------------------------------------------------------"
+                ui_print "----------------------------------------"
                 ui_print "${LOG_MSG_PREFIX}${LOG_MSG}"
-                ui_print "----------------------------------------------------------------------"
-            elif [ "$LOG_MSG_LEVEL" = "NONE" ]; then
+                ui_print "----------------------------------------"
+            elif [ "$LOG_MSG_LEVEL" = "-" ]; then
                 ui_print "$LOG_MSG"
             else
                 ui_print "${LOG_MSG_PREFIX}${LOG_MSG}"
@@ -167,25 +170,18 @@ logowl() {
 
 print_line() {
 
-    length=${1:-70}
+    length=${1:-40}
 
     line=$(printf "%-${length}s" | tr ' ' '-')
-    logowl "$line" "NONE"
+    logowl "$line" "-"
 }
 
-init_variables() {
+get_config_var() {
     key="$1"
     config_file="$2"
 
-    if [ -z "$key" ]; then
-        logowl "Key is NOT ordered! (1)" "ERROR"
-        return 1
-    fi
-
-    if [ ! -f "$config_file" ]; then
-        logowl "Config file $config_file does NOT exist (1)" "ERROR" >&2
-        return 1
-    fi
+    [ -z "$key" ] || [ -z "$config_file" ] && return 1
+    [ ! -f "$config_file" ] && return 2
 
     value=$(awk -v key="$key" '
         BEGIN {
@@ -219,10 +215,7 @@ init_variables() {
                             value = value "\n" remaining
                         }
                     }
-                    if (in_quote) {
-                        print "! Error: Unclosed quote for key " key > "/dev/stderr"
-                        exit 1
-                    }
+                    if (in_quote) exit 1
                 }
                 found = 1
             } else {
@@ -239,15 +232,13 @@ init_variables() {
         }
     ' "$config_file")
 
-    awk_exit_status=$?
-
-    case $awk_exit_status in
-        1)
-            logowl "Key '$key' does NOT exist in $config_file (5)" "WARN" >&2
+    awk_exit_state=$?
+    case $awk_exit_state in
+        1)  logowl "Key is NOT found or find unclosed quote! ($awk_exit_state)" "ERROR"
             return 5
             ;;
         0)  ;;
-        *)  logowl "Error occurred as processing key '$key' in $config_file ($awk_exit_status)" "WARN" >&2
+        *)  logowl "Unexpected error! ($awk_exit_state)" "ERROR"
             return 6
             ;;
     esac
@@ -263,7 +254,6 @@ init_variables() {
 }
 
 check_value_safety() {
-
     key="$1"
     value="$2"
 
@@ -275,6 +265,9 @@ check_value_safety() {
             logowl "Verified $key=$value (boolean)"
             return 0
         fi
+    else
+        logowl "Failed to verify key ($result_check_key)"
+        return 1
     fi
 
     check_param_safety "$value"
@@ -284,7 +277,7 @@ check_value_safety() {
         logowl "Verified $key=$value"
         return 0
     else
-        logowl "Safety check failed, key: $result_check_key, value: $result_check_value)"
+        logowl "Failed to verify value ($result_check_value)"
         return 1
     fi
 
@@ -316,18 +309,17 @@ check_param_safety() {
     return 0
 }
 
-verify_variables() {
-  
+verify_var() {
     config_var_name="$1"
     config_var_value="$2"
     validation_pattern="$3"
     default_value="${4:-}"
 
+    [ -z "$config_var_name" ] || [ -z "$config_var_value" ] || [ -z "$validation_pattern" ] && return 1    
+    
     script_var_name=$(echo "$config_var_name" | tr '[:lower:]' '[:upper:]')
 
-    if [ -z "$config_var_name" ] || [ -z "$config_var_value" ] || [ -z "$validation_pattern" ]; then
-        return 1    
-    elif echo "$config_var_value" | grep -qE "$validation_pattern"; then
+    if echo "$config_var_value" | grep -qE "$validation_pattern"; then
         export "$script_var_name"="$config_var_value"
         logowl "Set $script_var_name=$config_var_value" "TIPS"
         return 0
@@ -342,8 +334,7 @@ verify_variables() {
     fi
 }
 
-update_config_value() {
-
+update_config_var() {
     key_name="$1"
     key_value="$2"
     file_path="$3"
@@ -380,12 +371,10 @@ show_system_info() {
 }
 
 file_compare() {
-
     file_a="$1"
     file_b="$2"
     
     [ -z "$file_a" ] || [ -z "$file_b" ] && return 2
-    
     [ ! -f "$file_a" ] || [ ! -f "$file_b" ] && return 3
     
     hash_file_a=$(sha256sum "$file_a" | awk '{print $1}')
@@ -398,12 +387,9 @@ file_compare() {
 
 abort_verify() {
 
-    if [ -n "$VERIFY_DIR" ] && [ -d "$VERIFY_DIR" ] && [ "$VERIFY_DIR" != "/" ]; then
-        rm -rf "$VERIFY_DIR"
-    fi
-    print_line
-    logowl "$1" "WARN"
-    abort "This zip may be corrupted or have been maliciously modified!"
+    [ -n "$VERIFY_DIR" ] && [ -d "$VERIFY_DIR" ] && [ "$VERIFY_DIR" != "/" ] && rm -rf "$VERIFY_DIR"
+    logowl "$1" "ERROR"
+    abort "This zip may be corrupted or has been maliciously modified!"
 
 }
 
@@ -439,33 +425,7 @@ extract() {
     if [ "$expected_hash" == "$calculated_hash" ]; then
         logowl "Verified $file" >&1
     else
-        abort_verify "Failed to verify $file"
-    fi
-}
-
-clean_old_logs() {
- 
-    log_dir="$1"
-    files_max="$2"
-    
-    if [ -z "$log_dir" ] || [ ! -d "$log_dir" ]; then
-        logowl "$log_dir is not found or is not a dir! (1)" "ERROR"
-        return 1
-    fi
-
-    if [ -z "$files_max" ]; then
-        files_max=30
-    fi
-
-    logowl "Current log dir: $log_dir"
-    files_count=$(ls -1 "$log_dir" | wc -l)
-    if [ "$files_count" -gt "$files_max" ]; then
-        logowl "Clear old logs ($files_count as max allowed $files_max)"
-        ls -1t "$log_dir" | tail -n +$((files_max + 1)) | while read -r file; do
-            rm -f "$log_dir/$file"
-        done
-    else
-        logowl "$files_count files in $log_dir (max allowed $files_max)"
+        abort_verify "Failed to verify file $file"
     fi
 }
 
@@ -496,13 +456,8 @@ clean_duplicate_items() {
 
     filed=$1
 
-    if [ -z "$filed" ]; then
-        logowl "File is NOT provided! (1)" "ERROR"
-        return 1
-    elif [ ! -f "$filed" ]; then
-        logowl "$filed does NOT exist or is NOT a file! (2)" "ERROR"
-        return 2
-    fi
+    [ -z "$filed" ] && return 1
+    [ ! -f "$filed" ] && return 2
 
     awk '!seen[$0]++' "$filed" > "${filed}.tmp"
     mv "${filed}.tmp" "$filed"
@@ -510,14 +465,26 @@ clean_duplicate_items() {
 
 }
 
-debug_get_prop() {
-
+check_before_resetprop() {
     prop_name=$1
+    prop_expect_value=$2
+    prop_current_value=$(resetprop "$prop_name")
 
-    if [ -z "$prop_name" ]; then
-        logowl "Property name does NOT exist! (1)" "WARN"
-        return 1
+    [ -z "$prop_current_value" ] || [ "$prop_current_value" = "$prop_expect_value" ] || resetprop "$prop_name" "$prop_expect_value"
+
+}
+
+find_keyword_before_resetprop() {
+    prop_name="$1"
+    prop_contains_keyword="$2"
+    prop_expect_value="$3"
+    prop_current_value=$(resetprop "$prop_name")
+
+    [ -z "$prop_current_value" ] || [ -z "$prop_contains_keyword" ] || [ -z "$prop_expect_value" ] && return 1
+    
+    if echo "$prop_current_value" | grep -q "$prop_contains_keyword"; then
+        resetprop "$prop_name" "$prop_expect_value"
+        return 0
     fi
-    logowl "$prop_name=$(getprop "$prop_name")"
-    return 0
+
 }
